+++ pyprojectsort/__init__.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from .__version__ import __version__
 from .main import reformat_pyproject
 
-__all__ = ("__version__", "reformat_pyproject")
+__all__ = ("reformat_pyproject",)